case-data-master-engenharia-de-dados/requirements.txt
# Requisitos do Projeto
apache-airflow
boto3
cryptography
Faker
great-expectations
minio
numpy
pandas
psycopg2-binary
pyspark
python-dotenv
SQLAlchemy
urllib3
